+++ test/dummy/app/controllers/concerns/uses_default_layout.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module UsesDefaultLayout
+  extend ActiveSupport::Concern
+
+  included do
+    layout "flat_pack_sidebar"
+  end
+end