+++ test/dummy/config/initializers/recording_studio.rb
@@ -4,6 +4,7 @@ RecordingStudio.configure do |config|
   config.recordable_types = [
     "Workspace",
     "Page",
+    "User",
     "ChatThread",
     "ChatMessage",
     "RecordingStudioAttachable::Attachment"